+++ src/simulator/schedule.c
@@ -102,7 +102,7 @@ static unsigned choose_thread(void)
  * @details Simulates a loop scheduler.
  */
 void schedule
-(const double *tasks, unsigned ntasks, unsigned _nthreads, unsigned scheduler)
+(const unsigned *tasks, unsigned ntasks, unsigned _nthreads, unsigned scheduler)
 {
 	unsigned n;
